things in the right place
